Backend/requirements.txt
grpcio==1.76.0;
grpcio-tools==1.76.0;
fastapi==0.120.0;
uvicorn==0.38.0;